Fix: getPriceFull forwarded pricing type
This fix is from Kasper's #172 pull request. This removes an if condition that will never be true.

if given asset has PRICINGTYPE__FORWARDED, the resolvePricingConfig function returns the pricingType of the asset the pricing is forwarded to. hence, the pricingType must never be equal to PRICINGTYPE__FORWARDED after the pricing config is resolved.